.up xóa vé

+++ BanVeMayBay/BanVeMayBay/Controllers/CustomerController.cs
@@ -157,6 +157,7 @@ namespace BanVeMayBay.Controllers
         }
         public ActionResult ListOderCus()
         {
+         
             user sessionUser = (user)Session[Common.CommonConstants.CUSTOMER_SESSION];
             var listOrder = db.orders.Where(m=>m.CusId == sessionUser.ID).OrderByDescending(m=>m.ID).ToList();
             return View("ListOderCus", listOrder);
@@ -168,22 +169,39 @@ namespace BanVeMayBay.Controllers
         }
         public ActionResult canelOrder(int OrderId)
         {
+           
+
             order morder = db.orders.Find(OrderId);
             var orderDetail = db.ordersdetails.Where(m => m.orderid == morder.ID).ToList();
             foreach (var item in orderDetail)
             {
                 var id = int.Parse(item.ticketId.ToString());
                 ticket ticket = db.tickets.Find(id);
-                ticket.Sold = ticket.Sold - item.quantity;
-                db.Entry(ticket).State = EntityState.Modified;
-                db.SaveChanges();
-                if (item == null)
+                DateTime ngaymuon = Convert.ToDateTime(
+                    morder.created_ate);
+                DateTime ngaytra = Convert.ToDateTime(ticket.departure_date);
+                TimeSpan Time = ngaytra - ngaymuon;
+                int TongSoNgay = Time.Days;
+               if(TongSoNgay >= 14)
                 {
-                    Message.set_flash("Error Cancel Order", "danger");
+                    ticket.Sold = ticket.Sold - item.quantity;
+                    db.Entry(ticket).State = EntityState.Modified;
+                    db.SaveChanges();
+                    if (item == null)
+                    {
+                        Message.set_flash("Error Cancel Order", "danger");
+                        return Redirect("~/tai-khoan");
+                    }
+                    db.ordersdetails.Remove(item);
+                    db.SaveChanges();
+                }
+                else
+                {
+                    Message.set_flash("Tickets cannot be canceled 14 days before flight date", "dangger");
                     return Redirect("~/tai-khoan");
                 }
-                  db.ordersdetails.Remove(item);
-                db.SaveChanges();
+                
+               
             }
                   
             db.orders.Remove(morder);
